Move wrapper under internal/
